feat: move to .tsx
